core/requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
amqp==5.0.6
apiview==1.3.25
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.0.post0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
Django==3.2.4
django-allauth==0.44.0
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-grappelli==2.14.4
django-import-export==2.5.0
django-rest-auth==0.9.5
django-taggit==1.4.0
django-templated-mail==1.1.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drf-writable-nested==0.6.3
et-xmlfile==1.1.0
export==0.1.2
extras==1.0.0
fcm-django==0.3.10
filelock==3.0.12
fixtures==3.0.0
fuzzywuzzy==0.18.0
greenlet==1.1.0
httpie==2.4.0
httpio==0.3.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
lettuce==0.2.23
lettuce-webdriver==0.3.5
linecache2==1.0.0
MarkupPy==1.14
MarkupSafe==2.0.1
mock==4.0.3
multidict==5.1.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
pbr==5.5.1
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pycparser==2.20
pyfcm==1.4.9
Pygments==2.9.0
PyJWT==2.1.0
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-mimeparse==1.6.0
python-subunit==1.4.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
SQLAlchemy==1.4.14
sqlparse==0.4.1
sure==1.4.11
tablib==3.0.0
telebot==0.0.4
testtools==2.4.0
traceback2==1.4.0
typing-extensions==3.10.0.0
unittest2==1.1.0
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
virtualenv==20.4.4
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
